first example of timed mode working for first event (selecting chancellor) successfully implemented, now needs about 10 more..
